refactor: npm으로 배포된 모달 패키지를 사용하도록 로직 수정
