+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 
 import { PreviousPathProvider } from './utils/context/PreviousPathContext';
-import { ModalProvider } from './components/common/Modal/ModalContext';
+import { ModalProvider } from './utils/context/ModalContext';
 import { CardsProvider } from './domain/context/CardsContext';
 
 const App = () => {
