+++ src/components/CardCompanySelectBottomSheet/index.tsx
@@ -1,13 +1,16 @@
-import CardCompanyButton from '../CardCompanyButton';
-import Modal from '../common/Modal';
+import { Modal } from '@suyoungj/react-simple-modal-ts';
 
+import CardCompanyButton from '../CardCompanyButton';
 import { CARD_COMPANIES } from '../../domain/constants';
+import { useModalContext } from '../../utils/context/ModalContext';
 
 import styles from './cardCompanySelectBottomSheet.module.css';
 
 const CardCompanySelectBottomSheet = () => {
+  const { isModalOpen, closeModal } = useModalContext();
+
   return (
-    <Modal>
+    <Modal isModalOpen={isModalOpen} onCloseModal={closeModal}>
       <div className={styles.container}>
         <div className={styles.companyContainer}>
           {CARD_COMPANIES.map((cardCompany) => (
