+++ src/components/CardRegisterForm/index.tsx
@@ -1,16 +1,19 @@
+import { useEffect } from 'react';
+
 import Card from '../Card';
 import Input from '../common/Input';
 import Tooltip from '../Tooltip';
 import TooltipButton from '../TooltipButton';
+import CardCompanySelectBottomSheet from '../CardCompanySelectBottomSheet';
 
 import useCardRegisterForm from './useCardRegisterForm';
-import { useModalContext } from '../common/Modal/ModalContext';
+import { useModalContext } from '../../utils/context/ModalContext';
 import { CARD_NUMBER_INPUT_PLACEHOLDER } from '../../domain/constants';
 
 import styles from './cardRegisterForm.module.css';
 
 const CardRegisterForm = () => {
-  const { openModal } = useModalContext();
+  const { isModalOpen, openModal } = useModalContext();
   const {
     cardCompany,
     cardNumber1,
@@ -31,6 +34,16 @@ const CardRegisterForm = () => {
     handleSubmit,
     handleFormChange,
   } = useCardRegisterForm();
+
+  useEffect(() => {
+    openModal();
+  }, []);
+
+  useEffect(() => {
+    if (isModalOpen === false) {
+      cardNumberInputRef.current?.focus();
+    }
+  }, [isModalOpen]);
 
   return (
     <>
@@ -209,6 +222,7 @@ const CardRegisterForm = () => {
           <button className={styles.submitButton}>다음</button>
         )}
       </form>
+      <CardCompanySelectBottomSheet />
     </>
   );
 };
